Update baalStakingEligibility.json
Temporary adjustment to enable the hats app to accept user input for the min stake creation parameter. 

When the app encounters the `amountWithDecimals` display type, it tries to call the token contract to get number of decimals. But this module sets the token address (read from the Baal contract) when its deployed, so the app doesn't know which token contract to call before deployment.

Changing the display type to `default` gets around this by putting the onus on the user to know how many decimals the token has.